# src/swagger/config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trasactions API",
    "version": "1.0.0",
    "description":"This is a study API to practice the concept of transactions in prisma"
  },
  "tags": [
    {
      "name": "User",
      "description": "Operations related to user management"
    },
    {
      "name": "Operations",
      "description": "Operations related to general functionality"
    }
  ],
  "paths": {
    "/user/find-all": {
      "get": {
        "summary": "Returns a list of users.",
        "description": "Optional extended description.",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "A JSON array of user names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "username": {
                        "type": "string"
                      },
                      "balance": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/create": {
      "post": {
        "summary": "Create a new user.",
        "description": "Optional extended description.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "balance"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "username": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operations/deposit": {
      "get": {
        "summary": "Makes a deposit of value into a user's balance.",
        "description": "Makes a deposit of value into a user's balance",
        "tags": [
          "Operations"
        ],
        "responses": {
          "200": {
            "description": "Deposit made successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Please provide from, to and amount and User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operations/transfer": {
      "post": {
        "summary": "Transfer values",
        "description": "Transfers values from one user to another",
        "tags": [
          "Operations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "senderUserId": {
                    "type": "number"
                  },
                  "receiverUserId": {
                    "type": "number"
                  },
                  "value": {
                    "type": "number"
                  }
                },
                "required": [
                  "senderUserId",
                  "receiverUserId",
                  "value"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Please enter the source user and destination user Source user not found and Sender user or received user is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operations/statement/:userId": {
      "get": {
        "summary": "Makes a deposit of value into a user's balance.",
        "description": "Makes a deposit of value into a user's balance",
        "tags": [
          "Operations"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Statement retrieved successfully.",
            "content": {
              "application/json": {
                "example": {
                  "id": "number",
                  "userName": "number",
                  "balance": "number",
                  "sentTransfers": [
                    {
                      "id": "number",
                      "value": "number",
                      "createdAt": "string",
                      "senderUserId": "number",
                      "receiverUserId": "number"
                    }
                  ],
                  "receivedTransfers": [
                    {
                      "id": "number",
                      "value": "number",
                      "createdAt": "string",
                      "senderUserId": "number",
                      "receiverUserId": "number"
                    }
                  ],
                  "deposits": [
                    {
                      "id": "number",
                      "user": "number",
                      "value": "number",
                      "createdAt": "string"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}